api/users/service.py
from rest_framework_simplejwt.tokens import RefreshToken


class GetLoginResponseService:
    @staticmethod
    def get_login_response(user, request):
        refresh = RefreshToken.for_user(user)
        data = {"refresh": str(refresh), "access": str(refresh.access_token)}
        print(refresh)
        return data
